misc: Changed ksyms and debug to use uintptr_t
Changed KSyms::get_symbol() and Debug::trace_stack() to use uintptr_t
instead of uint64_t. Additionally changed the arguement of trace_stack
to be void* instead of uint64_t* as well as the associated
comments/documentation to clarify that it takes a frame pointer as an
arguement.

+++ include/kernel/arch/x86_64/ksyms.h
@@ -20,10 +20,10 @@ namespace KSyms {
 	 * @brief Get the symbol name for a given address
 	 *
 	 * @param addr The address to get the symbol name for
-	 * @param sym_addr The address of the symbol
+	 * @param sym_addr The start address the symbol refers to
 	 * @return The symbol name, or nullptr if not found
 	 */
-	const char *get_symbol(void *addr, uint64_t *sym_addr);
+	const char *get_symbol(void *addr, uintptr_t *sym_addr);
 
 	/**
 	 * @brief Initialize the kernel symbol table
